# server/api/openapi.yaml
openapi: 3.0.0
info:
  title: Baby Corona
  description: |
    Baby Corona API |
    [Baby Corona website](https://babysitter-il.appspot.com/)
  termsOfService: http://swagger.io/terms/
  contact:
    email: imomerm@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0-oas3
servers:
- url: https://virtserver.swaggerhub.com/api
- url: http://virtserver.swaggerhub.com/api
...
paths:
...
  /session/countmatched:
    post:
      tags:
      - session
      summary: Count All matched sessions
      operationId: countMatchedSessions
      responses:
        "200":
          description: amount of all matched sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                x-content-type: application/json
        "404":
          description: No matched sessions
      x-swagger-router-controller: Session
  /session/counturgentpending:
    post:
      tags:
      - session
      summary: Count All urgent requests
      operationId: countUrgentPendingSessions
      responses:
        "200":
          description: amount of all urgent pending sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                x-content-type: application/json
        "404":
          description: No matched sessions
      x-swagger-router-controller: Session
  /session:
    post:
      tags:
      - session
      summary: Create a new session (from volunteer request usually)
      operationId: createSession
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Session'
      responses:
        "200":
          description: Session created
        "405":
          description: Invalid input
      x-swagger-router-controller: Session
  /session/delete:
    post:
      tags:
      - session
      summary: delete session
      operationId: deleteSession
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                sessionId:
                  type: string
      responses:
        "200":
          description: session deleted
        "405":
          description: Invalid input
      x-swagger-router-controller: Session
  /session/addrequest/{sessionId}:
    post:
      tags:
      - session
      summary: Add a new request from a volunteer to the session
      operationId: addRequestToSession
      parameters:
      - name: sessionId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                volunteerId:
                  type: string
      responses:
        "200":
          description: Request added created
        "404":
          description: Not found
        "405":
          description: Invalid input
      x-swagger-router-controller: Session
  /session/history/{userId}:
    post:
      tags:
      - session
      summary: Get all sessions of a specified user (and role)
      operationId: getAllSessionsByUser
      parameters:
      - name: userId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/body'
      responses:
        "200":
          description: All sessions of the user selected
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
                x-content-type: application/json
        "404":
          description: User not found
        "405":
          description: Invalid input
      x-swagger-router-controller: Session
  /session/getUpcomingSessions/{userId}:
    post:
      tags:
      - session
      summary: Get all approved upcoming sessions for volunteer
      operationId: getAllUpcomingApprovedSessionsByVolunteer
      parameters:
      - name: userId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: All approved future seessions for volunteer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
                x-content-type: application/json
        "404":
          description: No sessions found
      x-swagger-router-controller: Session
  /session/getUpcomingNotYetApprovedSessions:
    post:
      tags:
      - session
      summary: Get all not yet approved upcoming sessions
      operationId: getAllUpcomingNotYetApprovedSessions
      responses:
        "200":
          description: All not yet approved future seessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
                x-content-type: application/json
        "404":
          description: No sessions found
      x-swagger-router-controller: Session
  /session/getAllFilledSessions:
    post:
      tags:
      - session
      summary: Get all approved  sessions
      operationId: getAllFilledSessions
      responses:
        "200":
          description: All approved  seessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
                x-content-type: application/json
        "404":
          description: No sessions found
      x-swagger-router-controller: Session    
  /session/getNotYetApprovedSessions/{userId}:
    post:
      tags:
      - session
      summary: Get all not yet approved upcoming sessions for volunteer
      operationId: getAllUpcomingNotYetApprovedSessionsByVolunteer
      parameters:
      - name: userId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: All not yet approved future seessions for volunteer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
                x-content-type: application/json
        "404":
          description: No sessions found
      x-swagger-router-controller: Session
  /session/notAssigned:
    post:
      tags:
      - session
      summary: Get all not yet assigned upcoming sessions 
      operationId: getAllNotAssignedSessions
      responses:
        "200":
          description: gets all not yet assigned future seessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
                x-content-type: application/json
        "404":
          description: No sessions found
      x-swagger-router-controller: Session
  /session/getAvailableSessions/{userId}:
    post:
      tags:
      - session
      summary: Get all not yet approved upcoming sessions for volunteer
      operationId: getAvailableSessions
      parameters:
      - name: userId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: All available sessions near the specified volunteer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
                x-content-type: application/json
        "404":
          description: No sessions found
      x-swagger-router-controller: Session
  /session/{doctorId}:
    post:
      tags:
      - session
      summary: Get all sessions by a doctor's ID, add isFilled true/false.
      operationId: getDoctorSessions
      parameters:
      - name: doctorId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                isFilled:
                  type: boolean
      responses:
        "200":
          description: All sessions of the doctor with the filter applied.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
                x-content-type: application/json
        "404":
          description: No sessions found.
      x-swagger-router-controller: Session
  /session/approve/{sessionId}:
    post:
      tags:
      - session
      summary: approve a session to be filled by volunteer
      operationId: approveSession
      parameters:
      - name: sessionId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                volunteerId:
                  type: string
      responses:
        "200":
          description: Session approved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
                x-content-type: application/json
        "404":
          description: Error while approving sessions.
      x-swagger-router-controller: Session
...
  /hamal:
    post:
      tags:
      - hamal
      summary: Create new hamal user
      operationId: createHamalUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Hamal'
      responses:
        "200":
          description: Hamal user created
        "405":
          description: Invalid input
      x-swagger-router-controller: Hamal
  /hamal/all:
    post:
      tags:
      - hamal
      summary: Get all hamal users
      operationId: getAllHamalUsers
      responses:
        "200":
          description: All hamal users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Hamal'
                x-content-type: application/json
        "404":
          description: No hamal users found
      x-swagger-router-controller: Hamal
  /hamal/approve-reject/{userId}:
    post:
      tags:
      - hamal
      summary: Approve or reject a user (doctor / volunteer)
      operationId: approveOrRejectUser
      parameters:
      - name: userId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/body_1'
      responses:
        "200":
          description: User has been approved/rejected & mail has been sent
        "404":
          description: User not found
        "405":
          description: Invalid input
      x-swagger-router-controller: Hamal
...
components:
  schemas:
    Volunteer:
      required:
      - birthday
      - email
      - tz
      type: object
      properties:
        id:
          type: string
          description: MongoID
        firstName:
          type: string
        lastName:
          type: string
        tz:
          type: string
          description: Teudat Zehut
        birthday:
          type: string
          format: date-time
        email:
          type: string
          format: email
        phone:
          type: string
        institute:
          type: string
          description: Where does the volunteer study?
        profession:
          type: string
          description: What does the volunteer study / work (if not student)
        facebook:
          type: string
          description: Link to the volunteer's Facebook profile
        city:
          type: string
        address:
          type: string
        pos:
          type: object
          properties:
            lat:
              type: number
            lng:
              type: number
        photo:
          type: string
          description: Link to the user's photo, hosted on a photo server/cloud
        isFemale:
          type: boolean
        hobbies:
          type: array
          items:
            type: string
        notes:
          type: string
      example:
        birthday: 2000-01-23T04:56:07.000+00:00
        profession: profession
        address: address
        notes: notes
        city: city
        tz: tz
        facebook: http://example.com/aeiou
        photo: photo
        phone: phone
        hobbies: hobbies
        name: name
        institute: institute
        id: id
        email: ""
      xml:
        name: Volunteer
    Doctor:
      required:
      - email
      type: object
      properties:
        _id:
          type: string
          description: MongoID
        firstName:
          type: string
        lastName:
          type: string
        tz:
          type: string
          description: Teudat Zehut
        email:
          type: string
          format: email
        phone:
          type: string
        institute:
          type: string
          description: Doctor's workplace
        profession:
          type: string
          description: Doctor's job / role
        city:
          type: string
        address:
          type: string
        pos:
          type: object
          properties:
            lat:
              type: number
            lng:
              type: number
        hobbies:
          type: array
          items:
            type: string
        notes:
          type: string
        children:
          type: array
          items:
            $ref: '#/components/schemas/Doctor_children'
      example:
        profession: profession
        address: address
        notes: notes
        phone: phone
        city: city
        hobbies: hobbies
        children:
        - isFemale: true
          age: 0
        - isFemale: true
          age: 0
        tz: tz
        name: name
        institute: institute
        id: id
        email: ""
      xml:
        name: Doctor
    Session:
      type: object
      properties:
        _id:
          type: string
          description: MongoID
        doctor_id:
          type: string
        timeRequested:
          type: string
          format: date-time
        timeApproved:
          type: string
          format: date-time
        requests:
          type: array
          description: List of all volunteers who requested this session
          items:
            type: string
        filledBy:
          type: string
        didHappen:
          type: boolean
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        tasks:
          type: array
          items:
            type: string
        recurring:
          type: string
          enum:
          - once
          - daily
          - weekly
          - monthly
        contact:
          type: object
          properties:
            name:
              type: string
            phone:
              type: string
      example:
        doctor:
          profession: profession
          address: address
          notes: notes
          phone: phone
          city: city
          hobbies: hobbies
          children:
          - isFemale: true
            age: 0
          - isFemale: true
            age: 0
          tz: tz
          name: name
          institute: institute
          id: id
          email: ""
        timeRequested: 2000-01-23T04:56:07.000+00:00
        timeApproved: 2000-01-23T04:56:07.000+00:00
        recurring: once
        startTime: 2000-01-23T04:56:07.000+00:00
        id: id
        requests:
        - birthday: 2000-01-23T04:56:07.000+00:00
          profession: profession
          address: address
          notes: notes
          city: city
          tz: tz
          facebook: http://example.com/aeiou
          photo: photo
          phone: phone
          hobbies: hobbies
          name: name
          institute: institute
          id: id
          email: ""
        - birthday: 2000-01-23T04:56:07.000+00:00
          profession: profession
          address: address
          notes: notes
          city: city
          tz: tz
          facebook: http://example.com/aeiou
          photo: photo
          phone: phone
          hobbies: hobbies
          name: name
          institute: institute
          id: id
          email: ""
        endTime: 2000-01-23T04:56:07.000+00:00
        tasks:
        - taskName: taskName
          taskSelected: true
        - taskName: taskName
          taskSelected: true
        didHappen: true
        contact: 054-4445555
      xml:
        name: Session
    Hamal:
      type: object
      properties:
        _id:
          type: string
          description: MongoID
        firstName:
          type: string
        lastName:
          type: string
        tz:
          type: string
          description: Teudat zehut
        password:
          type: string
        doctorsApproved:
          type: array
          items:
            $ref: '#/components/schemas/Doctor'
        volunteersApproved:
          type: array
          items:
            $ref: '#/components/schemas/Volunteer'
      description: User in hamal
      example:
        password: password
        volunteersApproved:
        - birthday: 2000-01-23T04:56:07.000+00:00
          profession: profession
          address: address
          notes: notes
          city: city
          tz: tz
          facebook: http://example.com/aeiou
          photo: photo
          phone: phone
          hobbies: hobbies
          name: name
          institute: institute
          id: id
          email: ""
        - birthday: 2000-01-23T04:56:07.000+00:00
          profession: profession
          address: address
          notes: notes
          city: city
          tz: tz
          facebook: http://example.com/aeiou
          photo: photo
          phone: phone
          hobbies: hobbies
          name: name
          institute: institute
          id: id
          email: ""
        tz: tz
        name: name
        id: id
        doctorsApproved:
        - profession: profession
          address: address
          notes: notes
          phone: phone
          city: city
          hobbies: hobbies
          children:
          - isFemale: true
            age: 0
          - isFemale: true
            age: 0
          tz: tz
          name: name
          institute: institute
          id: id
          email: ""
        - profession: profession
          address: address
          notes: notes
          phone: phone
          city: city
          hobbies: hobbies
          children:
          - isFemale: true
            age: 0
          - isFemale: true
            age: 0
          tz: tz
          name: name
          institute: institute
          id: id
          email: ""
      xml:
        name: Hamal
...
    body:
      type: object
      properties:
        role:
          type: string
          enum:
          - volunteer
          - doctor
    body_1:
      type: object
      properties:
        isApproved:
          type: boolean
        hamalUserId:
          type: string
        role:
          type: string
          enum:
          - volunteer
          - doctor
    Doctor_children:
      type: object
      properties:
        age:
          type: integer
        isFemale:
          type: boolean
      example:
        isFemale: true
        age: 0
...